+++ src/main/java/com/prgrms/boardjpa/user/UserService.java
@@ -3,7 +3,8 @@ package com.prgrms.boardjpa.user;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.boardjpa.post.NotExistException;
+import com.prgrms.boardjpa.commons.exception.NotExistException;
+import com.prgrms.boardjpa.user.domain.User;
 import com.prgrms.boardjpa.user.dto.UserDto;
 
 @Service
