+++ src/main/java/com/prgrms/boardjpa/user/domain/Hobby.java
@@ -1,4 +1,4 @@
-package com.prgrms.boardjpa.user;
+package com.prgrms.boardjpa.user.domain;
 
 public enum Hobby {
 	MOVIE,
